requirements.txt
mpmath==1.3.0
networkx==3.2.1
numpy~=1.0
pillow==10.2.0
pandas
wandb
omegaconf
colorama
tqdm
opencv-python
matplotlib
lightning
imageio
einops>=0.7
wandb-osh
diffusers
datasets
internetarchive
pytorchvideo
scipy
rotary_embedding_torch
torchmetrics[image]==0.11.4
timm
pyiqa==0.1.10
transformers
decord
ninja
imageio
hydra-core
accelerate
torchsde
colorlog
scikit-learn
plyfile
calflops
openai-clip
wandb[media]
deepspeed